ContentPage: Split Content in its own component
Called ContentLightbox because it will be part of a lightbox.

https://phabricator.endlessm.com/T31806
